deepfake-detection/requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python
facenet-pytorch
pyyaml
scikit-learn
tqdm
matplotlib
Pillow
